split no dependency funcs to utils
