+++ internal/repositories/files.go
@@ -1,11 +1,10 @@
 package repositories
 
 import (
-	"errors"
-	"os"
 	"path/filepath"
 	"receipt-wrangler/api/internal/models"
 	"receipt-wrangler/api/internal/simpleutils"
+	"receipt-wrangler/api/internal/utils"
 )
 
 func BuildFilePath(receiptId string, receiptImageId string, receiptImageFileName string) (string, error) {
@@ -53,60 +52,16 @@ func BuildGroupPath(groupId uint, alternateGroupName string) (string, error) {
 	return groupPath, nil
 }
 
-func WriteFile(path string, data []byte) error {
-	// TODO: Fix perms
-	err := os.WriteFile(path, data, 777)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func ReadFile(path string) ([]byte, error) {
-	bytes, err := os.ReadFile(path)
-	if err != nil {
-		return nil, nil
-	}
-
-	return bytes, nil
-}
-
 func GetBytesForFileData(fileData models.FileData) ([]byte, error) {
 	path, err := BuildFilePath(simpleutils.UintToString(fileData.ReceiptId), simpleutils.UintToString(fileData.ID), fileData.Name)
 	if err != nil {
 		return nil, err
 	}
 
-	bytes, err := ReadFile(path)
+	bytes, err := utils.ReadFile(path)
 	if err != nil {
 		return nil, err
 	}
 
 	return bytes, nil
-}
-
-func DirectoryExists(dir string, createIfNotExist bool) error {
-	_, err := os.Stat(dir)
-	if errors.Is(err, os.ErrNotExist) && createIfNotExist {
-		err = MakeDirectory(dir)
-		if err != nil {
-			return err
-		}
-	}
-
-	if !errors.Is(err, os.ErrNotExist) {
-		return err
-	}
-
-	return nil
-}
-
-func MakeDirectory(dir string) error {
-	err := os.Mkdir(dir, os.ModePerm)
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
